Cleaned up files
