+++ main.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Mon Nov  4 18:46:08 2024
+
+@author: Luka
+"""
+
+from gui_logic import Timesheet
+
+
+# Testing GUI
+if __name__ == "__main__":
+    # gui_constants.DEBUG = True
+    app = Timesheet()
